pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "htan-model-mappings"
version = "0.1.0"
description = "SSSOM mappings between HTAN data models"
requires-python = ">=3.10"
dependencies = [
    "sssom>=0.4.0",
    "pyyaml>=6.0",
    "linkml-runtime>=1.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4.0",
]
